# src/docs/api.yaml
openapi: 3.0.3
info:
  title: Photo Clustering & PDF Export API
  version: 1.0.0
  description: >
    사용자가 사진을 업로드하고, 장소별로 클러스터링·수정한 뒤
    각 장소를 1페이지 PDF로 출력하는 서비스의 백엔드 API.
servers:
  - url: http://localhost:8000/
paths:
...
  /clusters/{clusterId}:
    patch:
      summary: 클러스터 이름/순서 수정
      parameters:
        - in: path
          name: clusterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterUpdate"
            example:
              name: "대원지하차도 분리대(우측)"
              order_index: 2
      responses:
        "200":
          description: 수정된 클러스터
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterDetail"
  /clusters/{clusterId}/photos/{photoId}:
    delete:
      summary: 클러스터에서 사진 제거 (삭제)
      parameters:
        - in: path
          name: clusterId
          required: true
          schema:
            type: integer
        - in: path
          name: photoId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: 성공 (본문 없음)
  /clusters/{clusterId}/move-photo:
    post:
      summary: 사진을 다른 클러스터로 이동
      description: |
        예: 장소2의 9번 사진을 장소3으로 이동
        - src: /clusters/{clusterId}
        - body.target_cluster_id: 3
      parameters:
        - in: path
          name: clusterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovePhotoRequest"
            example:
              photo_id: 9
              target_cluster_id: 11
      responses:
        "204":
          description: 성공 (본문 없음)
...
components:
  schemas:
...
    ClusterUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
        order_index:
          type: integer
          nullable: true
      required: []
    ClusterDetail:
      type: object
      properties:
        id:
          type: integer
        session_id:
          type: integer
        name:
          type: string
          nullable: true
        order_index:
          type: integer
        photos:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              original_filename:
                type: string
    MovePhotoRequest:
      type: object
      properties:
        photo_id:
          type: integer
        target_cluster_id:
          type: integer
      required: [photo_id, target_cluster_id]
...